requirements.txt
psycopg2-binary
flask
flask_cors
flask_sqlalchemy
sqlalchemy
Flask-Migrate
werkzeug
PyJWT
